refactor/#63: sns event 생성 시 파일, 썸네일 업로드하는 use case 선언 및 구현
- UploadFileAndThumbnail 인터페이스에 파일을 생성하고, 업로드하고, 썸네일을 업로드하는 로직 리팩토링
- 파일 변환 port, html 생성 port를 구현하여 책임 분리
