+++ src/main/java/com/haru/api/snsEvent/application/port/in/UploadFileAndThumbnail.java
@@ -1,0 +1,9 @@
+package com.haru.api.snsEvent.application.port.in;
+
+import com.haru.api.snsEvent.domain.SnsEvent;
+
+public interface UploadFileAndThumbnail {
+
+    String createAndUploadListFileAndThumbnail(SnsEvent snsEvent);
+
+}
