+++ src/main/java/com/haru/api/snsEvent/infrastructure/adapter/ParticipantPersistenceAdapter.java
@@ -2,7 +2,6 @@ package com.haru.api.snsEvent.infrastructure.adapter;
 
 import com.haru.api.snsEvent.application.port.out.ParticipantPort;
 import com.haru.api.snsEvent.domain.Participant;
-import com.haru.api.snsEvent.domain.SnsEvent;
 import com.haru.api.snsEvent.infrastructure.jpa.ParticipantJpaRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -16,7 +15,12 @@ public class ParticipantPersistenceAdapter implements ParticipantPort {
     private final ParticipantJpaRepository participantJpaRepository;
 
     @Override
-    public List<Participant> findAllBySnsEvent(SnsEvent foundSnsEvent) {
-        return participantJpaRepository.findAllBySnsEvent(foundSnsEvent);
+    public void saveAll(List<Participant> participants) {
+        participantJpaRepository.saveAll(participants);
+    }
+
+    @Override
+    public List<Participant> findAllBySnsEventId(Long snsEventId) {
+        return participantJpaRepository.findAllBySnsEventId(snsEventId);
     }
 }
