+++ src/main/java/com/haru/api/snsEvent/infrastructure/adapter/WinnerPersistenceAdapter.java
@@ -1,7 +1,6 @@
 package com.haru.api.snsEvent.infrastructure.adapter;
 
 import com.haru.api.snsEvent.application.port.out.WinnerPort;
-import com.haru.api.snsEvent.domain.SnsEvent;
 import com.haru.api.snsEvent.domain.Winner;
 import com.haru.api.snsEvent.infrastructure.jpa.WinnerJpaRepository;
 import lombok.RequiredArgsConstructor;
@@ -16,7 +15,12 @@ public class WinnerPersistenceAdapter implements WinnerPort {
     private final WinnerJpaRepository winnerJpaRepository;
 
     @Override
-    public List<Winner> findAllBySnsEvent(SnsEvent foundSnsEvent) {
-        return winnerJpaRepository.findAllBySnsEvent(foundSnsEvent);
+    public void saveAll(List<Winner> winners) {
+        winnerJpaRepository.saveAll(winners);
+    }
+
+    @Override
+    public List<Winner> findAllBySnsEventId(Long snsEventId) {
+        return winnerJpaRepository.findAllBySnsEventId(snsEventId);
     }
 }
