+++ src/main/java/com/haru/api/snsEvent/infrastructure/jpa/ParticipantJpaRepository.java
@@ -1,13 +1,12 @@
 package com.haru.api.snsEvent.infrastructure.jpa;
 
 import com.haru.api.snsEvent.domain.Participant;
-import com.haru.api.snsEvent.domain.SnsEvent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
 public interface ParticipantJpaRepository extends JpaRepository<Participant, Long> {
 
-    List<Participant> findAllBySnsEvent(SnsEvent foundSnsEvent);
+    List<Participant> findAllBySnsEventId(Long snsEventId);
 
 }
