+++ src/main/java/com/haru/api/snsEvent/infrastructure/jpa/WinnerJpaRepository.java
@@ -1,6 +1,5 @@
 package com.haru.api.snsEvent.infrastructure.jpa;
 
-import com.haru.api.snsEvent.domain.SnsEvent;
 import com.haru.api.snsEvent.domain.Winner;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -10,6 +9,6 @@ import java.util.List;
 @Repository
 public interface WinnerJpaRepository extends JpaRepository<Winner, Long> {
 
-    List<Winner> findAllBySnsEvent(SnsEvent foundSnsEvent);
+    List<Winner> findAllBySnsEventId(Long snsEventId);
 
 }
